refactor(ast): move non-node helpers to the @kubb/ast/utils subpath
resolveGroupType and extractStringsFromNodes return plain values, not
nodes, so they belong on the node-tree-free @kubb/ast/utils entry next to
the other string helpers. Move both there (plus the ParamGroupType and
BuildGroupArgs types) and drop them from the main @kubb/ast barrel.

Update the first-party consumers (@kubb/core FileProcessor and
@kubb/parser-md) to import extractStringsFromNodes from @kubb/ast/utils.
