+++ packages/ast/src/utils/index.ts
@@ -1,7 +1,10 @@
 import { isIdentifier, pascalCase, singleQuote } from '@internals/utils'
 import { INDENT } from '../constants.ts'
+import type { CodeNode, OperationNode, ParameterNode } from '../nodes/index.ts'
+import type { OperationParamsResolver, ParamGroupType } from './ast.ts'
 
 export { isValidVarName } from '@internals/utils'
+export type { BuildGroupArgs, ParamGroupType } from './ast.ts'
 
 /**
  * Strips a single matching pair of `"..."`, `'...'`, or `` `...` `` from both ends of `text`.
@@ -294,4 +297,66 @@ export function enumPropName(parentName: string | null | undefined, propName: st
 export function findDiscriminator(mapping: Record<string, string> | undefined, ref: string | undefined): string | null {
   if (!mapping || !ref) return null
   return Object.entries(mapping).find(([, value]) => value === ref)?.[0] ?? null
+}
+
+/**
+ * Extracts all string content from a `CodeNode` tree recursively.
+ *
+ * Collects text node values, identifier references in string fields (`params`, `generics`, `returnType`, `type`),
+ * and nested node content. Used to build the full source string for import filtering.
+ */
+export function extractStringsFromNodes(nodes: Array<CodeNode> | undefined): string {
+  if (!nodes?.length) return ''
+  return nodes
+    .map((node) => {
+      // Backward-compat: compiled plugins may still pass bare strings at runtime
+      if (typeof node === 'string') return node as string
+      if (node.kind === 'Text') return node.value
+      if (node.kind === 'Break') return ''
+      if (node.kind === 'Jsx') return node.value
+
+      const parts: Array<string> = []
+
+      if ('params' in node && node.params) parts.push(node.params)
+      if ('generics' in node && node.generics) parts.push(Array.isArray(node.generics) ? node.generics.join(', ') : node.generics)
+      if ('returnType' in node && node.returnType) parts.push(node.returnType)
+      if ('type' in node && typeof node.type === 'string') parts.push(node.type)
+
+      const nested = extractStringsFromNodes(node.nodes)
+
+      if (nested) parts.push(nested)
+
+      return parts.join('\n')
+    })
+    .filter(Boolean)
+    .join('\n')
+}
+
+/**
+ * Derives a {@link ParamGroupType} for a query or header group from the resolver.
+ *
+ * Returns `null` when there is no resolver, no params, or the group name equals the
+ * individual param name (so there is no real group to emit).
+ */
+export function resolveGroupType({
+  node,
+  params,
+  group,
+  resolver,
+}: {
+  node: OperationNode
+  params: Array<ParameterNode>
+  group: 'query' | 'header'
+  resolver: OperationParamsResolver | undefined
+}): ParamGroupType | null {
+  if (!resolver || !params.length) {
+    return null
+  }
+  const firstParam = params[0]!
+  const groupMethod = group === 'query' ? resolver.resolveQueryParamsName : resolver.resolveHeaderParamsName
+  const groupName = groupMethod.call(resolver, node, firstParam)
+  if (groupName === resolver.resolveParamName(node, firstParam)) {
+    return null
+  }
+  return { type: groupName, optional: params.every((p) => !p.required) }
 }
