+++ packages/core/src/FileProcessor.ts
@@ -1,6 +1,6 @@
 import { AsyncEventEmitter } from '@internals/utils'
 import type { CodeNode, FileNode } from '@kubb/ast'
-import { extractStringsFromNodes } from '@kubb/ast'
+import { extractStringsFromNodes } from '@kubb/ast/utils'
 import { STREAM_FLUSH_EVERY } from './constants.ts'
 import type { Storage } from './createStorage.ts'
 import type { Parser } from './defineParser.ts'
