+++ packages/parser-md/package.json
@@ -53,6 +53,7 @@
     "typecheck": "tsc -p ./tsconfig.json --noEmit --emitDeclarationOnly false"
   },
   "dependencies": {
+    "@kubb/ast": "workspace:*",
     "@kubb/core": "workspace:*",
     "yaml": "catalog:"
   },
